minior update on cbi_tool.cpp
- g++ will treat pow(obs_P_in_hour, Q_s) and pow(DOC_ratio, Q_n)
  as double but not clang++. thus, some casting must be done to make
  them as double.

+++ src/cpp/cbi_tool.h
@@ -392,11 +392,13 @@ public:
             {
                 Q_n = 1.124; // default, to ensure the mu is decreasing pattern as a function of D/C
                 //Cd = P / (D / C) ^ n
-                Q_cd = obs_P_in_hour / max(0.0001, pow(DOC_ratio, Q_n));
+                // Peiheng, 02/22/22, g++ will treat pow(DOC_ratio, Q_n) as double but not clang++
+                Q_cd = obs_P_in_hour / max(0.0001, (double)pow(DOC_ratio, Q_n));
             }
 
             //vc / vt2 - 1 = cp * (P)^s, --> cp = [ vc/vt2 - 1] / (P^s)  // assume s is fixed
-            Q_cp = (FD_vcutoff / max(0.0001f, t2_speed) - 1.0) / max(0.00001, pow(obs_P_in_hour, Q_s));
+            // Peiheng, 02/22/22, g++ will treat pow(obs_P_in_hour, Q_s) as double but not clang++
+            Q_cp = (FD_vcutoff / max(0.0001f, t2_speed) - 1.0) / max(0.00001, (double)pow(obs_P_in_hour, Q_s));
             //backward derivation
         }
